Created new directories for interfaces and constants and moved Command interface and CommandList constant to their respective locations
