+++ src/constants/CommandList.ts
@@ -1,9 +1,9 @@
-//command class
-import { Command } from "./Command";
-import { DiscussionMenu } from "./commands/DiscussionMenu";
-import { RegisterDiscussionCourse } from "./commands/RegisterDiscussionCourse";
+//command interface
+import { Command } from "../interfaces/Command";
 
 //import command from its respective file
+import { DiscussionMenu } from "../commands/DiscussionMenu";
+import { RegisterDiscussionCourse } from "../commands/RegisterDiscussionCourse";
 
 //list of all the available commands
 export const CommandList: Command[] = [DiscussionMenu, RegisterDiscussionCourse];
