+++ src/interfaces/Command.ts
